Fix models namespace
