+++ src/Models/SubscriptionAgreement.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace Payavel\Subscription\Models;
 
 use Illuminate\Database\Eloquent\Model;
